Replace non-definitional EXIT with UNWIND
Adding definitional RETURN meant that there actually had to be a
unique local in each function to bind to, holding a specific instance
of a function that knew how to return from a specific function.  If
Rebol2's EXIT were to be accommodated as well, each function would need
two of these words.

So users of FUNCTION were asked to just say `return ()`.  Yet to better
deal with the concept of routines that weren't supposed to return a
value, PROCEDURE was introduced.  But when it needed an arity-0 word
to mean "end the procedure and return to the caller with no result",
the word LEAVE was chosen.

There was controversy over EXIT's role in the language.  The mechanism
by which RETURN and EXIT had given the illusion of working was that
natives would not intercept them...so they would return from the
nearest usermode function.  That was ended, so EXIT became more or
less useless, unless it was used with a /FROM refinement that would
explicitly say which function FRAME! it wanted to jump to.  This
could be used to write usermode implementations of function generators
that provided definitional RETURN.

However, the name EXIT for this wasn't particularly great.  And its
behavior was now more confusing...since the 0-arity version didn't
do anything anyone was likely to actually mean.

So this creates UNWIND to be the arity-1 form for targeting a specific
function frame to return from, and deprecates EXIT so that it gives
an informative message.  Eventually, EXIT will probably have the
behavior of what EXIT-REBOL does today...which is to give a guaranteed
termination of the interpreter without possibility of being caught
the way a QUIT is.