Improve Time Zone Handling in PICK and POKE
R3-Alpha had an unusual idea for when a date was manipulated into an
invalid state, such as to give it an invalid day of the month or an
invalid time.  It would "normalize" the date as if math had been
performed on it to send it to an invalid state, effectively rounding
the answer.  So requesting Feb 29th on a non-leap-year would round up
to March 1st...for instance.

Consensus on the forum was that errors were better, and more generally
that the system should be more correct about making the user experience
give back what was expected...e.g. asking for or setting the day or
time zone should change *just that element* without disturbing the
others.  This requires internal adjustments (e.g. if the time zone is
changed then the bits of the date have to be fixed up).

This begins the process of implementing that.