Introduce CSCAPE templating routine for simplifying bootstrap
R3-Alpha's "make prep" step centered around generating a large number
of derived .C and .H files.  The input to this generation process was
several Rebol tables, as well as scanning over the C source code of
the interpreter itself.

The make prep code was somewhat ad-hoc, and used Rebol JOIN and other
string operations to build up lines of C.  During the development of
Ren-C, the demands on this process became greater...so the already
somewhat difficult-to-follow code became more complicated than it
already was.

To begin taming the complexity, this introduces a new way to merge
Rebol variables into a template of C code.  A templating routine
called CSCAPE (for C escape, pronounced "sea-scape") is integrated
into the emitter.  Rebol escapes identified by ${...} are LOADed
and evaluated, and their result has TO-C-NAME run on them.  If $(...)
is used instead, then the result is run through UNSPACED (if a block)
or FORM (for all other types).

An added twist is the usage of capitalization in the expression, which
affects the result.  Use of all uppercase characters means the result
will be run through UPPERCASE, all lowercase means it will run through
LOWERCASE, and mixed casing (or if no letters are present) means it
will be left alone.  Hence:

    mod: {Super-Crypto+}
    n: 10

    >> cscape {int EXT_${MOD}_NUM = $(n);}
    == {int EXT_SUPER_CRYPTO_A_NUM = 10;}

    >> cscape {const char *s = "${Mod}";}
    == {const char *s = "Super_Crypto_A";}

This permits *significant* simplification in the emit process.  As a
first step, this integrates CSCAPE into the basic emit command, and
uses it in a few places to demonstrate it.

Because the LOAD only binds into the user context, this trick won't
work for function locals.  But so long as a Ren-C is used for bootstrap
(which it appears we are now committed to) then the FRAME! (or a word
in the frame) can be passed in via a /WHERE refinement, and it will
also look there for bindings.