Type check deferred arguments at the appropriate time
When argument deferral was introduced, type checking was not also
deferred.  This means an operation like `EITHER () = () [...] [...]`
would fail, because it would go through an intermediate step of
`EITHER ()`, and type check that, before getting to the need for a
next argument and re-entering a subframe to complete the enfix...
producing a LOGIC!, legal for the either condition.

This pointed to the general fact that an argument should not be
typechecked at the moment of fulfillment if it may wind up changing
by means of a frame re-entry.  It would have to be type-checked at
the moment the deferred parameter was fulfilled...and if no re-entry
happened to change the parameter, its deferred type checking could
not be forgotten about.

Since the situation only comes up between evaluations--before the
frame's cell would be used for anything else--this commit solves the
problem by remembering the parameter and refinement state in fields
of that cell.  Then, typechecking is moved out into a subroutine
which passes in that state.  It's not too invasive a fix, and doesn't
add any extra size to the frame.

(Note: it actually could allowing for f->deferred to be eliminated,
since the VAL_TYPE_RAW() of the f->cell can be checked...assuming
there is no other situation which could give rise to a REB_0 in the
frame's cell.)