Remove Extension-Words feature in favor of libRebol word access
Extensions needed to do comparisons and mentions of WORD! from C code.
This was done by listing the words they wanted to register in comment
blocks, then the prep process would generate the code to register
a set of REBSTR* series pointers into an array.

One mechanical problem with this is that those REBSTR* arrays were not
known to the GC.  Hence it was technically possible for the spellings
to be GC'd and then cause a crash.

But more generally, the libRebol API is focused on making it easy to
work with Rebol code in a direct format.  If someone wanted to cache a
GC-protected word value in an extension, they should be able to do
so fairly naturally with the API...then the API handles would (should)
be scoped to the lifetime of the extension.  (Today they would have to
be marked as unmanaged.)

In the ongoing interest of simplification and exercising the API,
clients who relied on the words list now use the API directly.  This
was %mod-crypt.c, %mod-locale.c, and %mod-odbc.c.