Update to build RIGOROUS with MSVC 2017 update (May 2018)
Ren-C tries to build with warnings enabled, and remedy problems instead
of disabling them...by promoting any non-disabled warnings to errors.
However, each new release of a compiler can add or change warnings, so
this has to be an ongoing process.

This updates the code to deal with new warnings in a just-released
update to MSVC.  The most significant one is related to the Spectre
vulnerability, where a warning tells you if the compiler would insert
code that would slow down your compiled result.

It is possible to rearrange code to not trigger this warning:

https://stackoverflow.com/questions/50399940/

However, as with other warnings, it's best not to do this in 3rd party
code (or code from R3-Alpha that may be dated and not necessary in the
future).  So this only changes some of the code to be compliant, while
disabling the warnings in things like %u-zlib or %aes.c.