Replace EMIT-ITEM/EMIT-LINE usages with CSCAPE EMIT
This pushes the build process to use more CSCAPE-based code, which
gives a better view of the templated output of the C being produced.

It introduces the new features of $[] to output a list with a delimiter
on each line (e.g. `$[fields];` can be used output a block that
contains a list of struct fields with a semicolon on each line) and
$() to output a list with a delimiter on all lines but the last
(e.g. `$(enum-items),` will put a comma on each line except the last.

This means that many "emit-as-you-go" bits of code have been changed
to "collect and then emit".  While somewhat slower, it is vastly
more clear to read.

Also changes the type table to not have an entry for 0.