Rethink TO behavior for ANY-ARRAY!, move complexity to MAKE
A long-discussed behavior has been to not automatically splice paths
when they are appended to BLOCK!s or GROUP!s.  This implements that
behavior:

    >> append copy [a b c] 'd/e/f
    == [a b c d/e/f]

However, BLOCK! and GROUP! continue to splice when APPENDed or INSERTed
or CHANGEed into another ANY-ARRAY! without an /ONLY refinement.  So
the new splicing rule is really only for PATH!--which now just won't
splice unless they are being spliced into *other paths*.

As a further step in the design, this makes it consistent when doing
an APPEND of things of different types, that it will act as if the
item being appended had first been converted to the target type.  This
helps reign in the TO matrix when there are multiple potential meanings,
by choosing the same interpretation for a plain APPEND as a TO.

e.g.

    block: copy [a b]
    m: make map! [c 10 d 20]
    append block m ;-- ??

Should the APPEND be adding one MAP!, or four items (c,10,d,20)...or
something else?  The concept in this commit is that whatever it did
do, it would be the same as:

    append block to block! m

Given that the desired behavior for the APPEND is to add one MAP!, that
would then suggest that TO BLOCK! of a MAP! under these rules would
need to be that MAP! in a single element block.

This commit pushes on that rule.  Among some of the changes, it affects
how TO BLOCK! of a TEXT! works to just putting the string into a block,
because:

    >> append copy [a b c] "d <e> #f"
    == [a b c "d <e> #f"]

If that had been [a b c d <e> #f] as an outcome, then TO BLOCK! would
have reasonably invoked the scanner on the string.  But for consistency
that is now needed to be moved to another operation.  (In that case,
MAKE of a BLOCK! from a TEXT! always did this, so that can be used.)