Change loop protocol: BREAK=>NULL, NeverRun=>BLANK!
Since loops returned the last result from their body evaluation, it
had seemed natural that if the body never ran...then the loop should
return NULL...much like a conditional that did not take its branch.
This avoided the synthesis of an arbitrary value, and allowed callers
of loops to recognize that a null result meant the body never ran.

But a more important distinction is whether or not a break happened.
For a time this was implemented by disallowing loop bodies from giving
falsey results, coercing the outcome to BAR!.  So a truthy result would
mean the loop ran at least once, a null result that it never ran, and
blank would mean a BREAK occurred.

The "Bar-ificiation" of blanks and falses, however, caused disruption
which was "truthifying" what the loop author had intended as a falsey
result.  This commit switches the situation such that BREAK signals
by returning NULL, and if a loop never runs it returns blank.  It is
then up to the user to decide if they wish to make the loop return a
non-blank thing to distinguish the cases that the loop did run.

This gives nearly the same level of control, albeit the user must be
a bit more complicit.