Add Update of Andreas Bolka's R3-Alpha ZeroMQ extension
This converts the R3-Alpha extension for ZeroMQ to Ren-C's extension
model, rewriting it to take advantage of mixing inlined script code
with API values in libRebol.  Original code from:

https://github.com/earl/r3-zmq

It also updates the ZeroMQ binding options and API names to version 4
from the version 2 it was originally targeting (in 2011).

There are three samples that are brought up to date as well:

* A hello world client and hello world server, demonstrating the most
  basic message passing over a socket.

* A weather server that randomly publishes made-up weather data for
  various zipcodes, and a client which subscribes to the data for only
  one of those zipcodes...which waits to gather 100 of them and then
  calculates the average of those 100 temperature readings.

* A workload distribution system, where one kind of process will take
  work items out of a list and perform them, another will gather the
  results of that processing, and another will "ventilate" tasks into
  the system to be done.

Some code (such as the zmq-poll()) is not yet tested, as there were no
usages in the samples.

At present it uses a mixture of internal API and libRebol, but it is
gearing up to be libRebol-based only.