DO/NEXT interface is now EVALUATE, empty GROUP! is void
This commit improves the interface by which a DO/NEXT is done.  Before,
each DO/NEXT would return the evaluative result (like DO) but allow
the updated position to be stored in a variable:

    >> do/next [1 + 2 10 + 20] 'pos
    == 3

    >> pos
    == [10 + 20]

This changes that to return the new position and *optionally* update
a variable with the output result.  Since this changes the contract
far from what DO offers, it gets the new name EVALUATE:

    >> evaluate/set [1 + 2 10 + 20] 'value
    == [10 + 20]

    >> value
    == 3

The optional update of the variable is key.  This means that in
successive evaluator steps, the output need not be overwritten.  This
deals with the problem of "residual invisibles", e.g.

    >> evaluate/set [1 + 2 elide (print "hi")] 'var
    == [elide (print "hi")]

    >> var
    == 3

    >> evaluate/set [elide (print "hi")] 'var
    ;-- null

    >> var
    == 3

With this dynamic in hand, it's possible to defer elided material so
that it runs along with what comes after it (if anything).  This
is preferable to running with the previous thing, which led to
undesirable behaviors like:

    >> any [1 + 2 elide (print "hi")]
    hi ;-- would have preferred NOT to run this
    == 3

This change also includes a migration so that during ordinary eval,
`()` is #[void] and not null, as is `do []`.  Generation of NULL is
done with the native NULL.  But other operations such as COMPOSE
are able to decide to treat `compose [a () b]` as being `[a b]`, as
opposed to void as `compose [a (print "hi") b]` would do.