JS-AWAITER and rebPromise() JavaScript features
This is a first cut at implementing a variant on JS-NATIVE, tentatively
called a JS-AWAITER.  What an awaiter does is that it simulates
a synchronous operation which is actually done through asynchronous
JavaScript calls.  Paralleling JS promises, it does this by not
returning to the caller just when the function body ends...rather, some
asynchronous event can later cause the return by calling resolve().

A simple JS-AWAITER implementing a timer could look like this:

    sleep: js-awaiter [
        {Sleep for the requested number of seconds}
        seconds [integer! decimal!]
    ]{
        setTimeout(
            resolve,
            1000 * rebUnboxDecimal(rebR(rebArg("seconds")))
        )
    }

While it would quickly finish running the body, the caller of the
SLEEP function wouldn't be returned to until JavaScript later called
the resolve() function.

Mechanically, awaiters cannot run synchronously on the JavaScript
call stack.  e.g. if you wrote:

    var x = rebUnboxInteger("sleep 5 | 10 + 20")
    console.log("x is " + x)

This wouldn't be able to work.  The setTimeout won't have a chance to
run if the code can't get up to the main GUI loop, and getting to the
GUI loop will involve pushing the stack forward past the console.log()
with a sum it should't have until after the sleep is finished.

To remedy this situation, the rebPromise() is introduced, which allows
calling JS-AWAITER functions, at the cost of not giving back a result
synchronously:

    var promise = rebPromise("sleep 5 | 10 + 20")
    promise.then(function (value) {
        console.log("x is " + rebUnboxInteger(rebR(value)))
    })

This is fairly standard in JavaScript, and once you are working with
promises you can use techniques like ASYNC/AWAIT:

https://javascript.info/async-await

Other changes in this commit include adding a rebTick() function to
help debugging, as well as factoring out the EnterApi() API check to
be put in by the %make-reb-lib.r script instead of manually in the
API definitions.

One issue about this commit is that it brings in some JavaScript
specific code under an #ifdef in the core %a-lib.c file to implement
the rebPromise().  It may be possible to put that functionality into
a MAKE-PROMISE native of some kind, so it can live in %mod-javascript.c
which will be something to look into.