Add prebuilt 32/64-bit MinGW ZeroMQ libs to %external/
This commit contains builds of ZeroMQ made manually on a local virtual
machine set up to mirror the conditions of the Travis CI containers
that cross-compile Windows builds on Linux.  If the build process were
adapted to use these pre-built static libraries, it could produce an
executable that would not require users to install any DLLs to use
ZeroMQ functionality.