IS/ISN'T Step 1: temp. remove = != <> EQUAL? NOT-EQUAL?
The ultimate goal of this commit is to eliminate the STRICT-EQUAL?
family of functions and operators, by not having a form of equality
that is "more equal than equal".  Instead, the lax form of equality
in default use by FIND, SWITCH and other operators would be called
"is-ness".  So "A" is not equal to "a", but `"A" IS "a"`.

This adds IS and ISN'T, while temporarily making the former lax equal
operations undefined.  However, the strict operators are still defined.
This gives people a chance to find all the `=` in their code and decide
if it should become a `==` or an IS (for example).

One difference between IS and ISN'T and previous lax equality is that
it does not consider different string types to match, nor does it
consider different word types to match.  Hence `<FOO> is <fOo>` but
`<FOO> isn't "FOO"`.  Integers and decimals continue to match within
a certain epsilon when using IS.

The next step in this process would be taken when all files have been
adapted for this change.  It would proceed to undefine the strict
equality operators, and have them retake `=`, `!=`, and `<>`.