Make FOR-NEXT/FOR-SKIP take word, introduce ITERATE
Rebol2/R3-Alpha had a FORALL construct, whose name did not obviously
distinguish it from FOREACH.  (What's the difference between doing
something "for all" elements in a block and "for each" element in a
block?)

However it was different in two big ways.  One was that it fetched the
series via word from a variable you passed in, and modified that
variable as it went--vs. taking in a variable name separate from what
was to be enumerated as FOR and FOR-EACH did.  This creates the
question of under what conditions that variable should be put back
to the initial series position.  It would do so in a THROW situation,
but not on an error, see:

https://github.com/rebol/rebol-issues/issues/2331

The other way in which it was different was that it visited each series
position, instead of each value.  It was also possible to change the
position during the iteration, so that it would NEXT whatever it was

In Ren-C, the name was changed to FOR-NEXT, which made more sense in
distinguishing the nature of the iteration.  But not taking a variable
to iterate made it odd for a "FOR" function, and also introduced the
question of what happened to the variable on loop ending.

This commit breaks those issues out by making FOR-NEXT take a variable
to set.  If you use the FOR-NEXT form, you don't have to worry about
what happens to the "original variable" in the iteration, because it
is a normal loop variable:

    for-next pos data [...]

Whatever you passed in as data was by value, not a word.  But FOR-NEXT
obeys the protocol of LIT-WORD! meaning to "use an existing binding".
So you can write:

    for-next 'data data [...]

This is the basis for a specialization called ITERATE, which does this
for you and provides the functionality of FORALL.  Decisions on what
this should do in response to throws or errors then becomes part of
that userspace construct.  For the moment, it does not do any restore
on either throws or errors...and those who are concerned about the
state of the iteration in such cases should use FOR-NEXT instead.