Fix latent Do_Any_Array_At_Throws(D_OUT, D_OUT) use
At one time, it was apparently safe to use the same value as the input
and the output to Do_Any_Array_At_Throws().  But when it was changed
to do Init_Void(D_OUT) in its implementation as part of the function
arguments, that changed...and depending on argument evaluation order,
it could try to extract a VAL_ARRAY() on a void...which was an error.