Immutable Paths, of minimum length 2
This changes PATH! to no longer be an ANY-SERIES! type.  This means it
cannot be operated on with series operations like APPEND, and it can't
be traversed to give new positions into paths with things like NEXT
or SKIP.  However, other operations like LENGTH OF, PICK, and FOR-EACH
continue to work.

Because there's no way to change them, ANY-PATH! is also immutable.
This means that a path can be checked at the time of its creation for
properties which cannot be violated by future mutations.

The new rules are:

* Paths must contain at least 2 elements.  The minimum path is `/`,
  which is the path form of [_ _]...two blanks.

* Paths may not directly contain other paths.  This prevents various
  ambiguity problems like [a/b c] and [a b/c] both being possible
  interpretations of a/b/c.  Indirect inclusion of other paths is
  legal, such as `a/(b/c)/d`

It is believed that this change eliminates harmful degrees of freedom,
without really sacrificing any behaviors people found that interesting.
Additionally, being able to promise paths are immutable open doors to
being able to create optimized representations of things like `/` or
`/a`, which could be packed into a single cell with no array node.

If one wishes to do surgery on path content, the way to do that is to
convert it to a BLOCK! or GROUP! and use series operations on that,
and then change it back to a PATH!.

(Note: Any series embedded in a path, e.g. the GROUP! in `a/(b c d)/e`,
is not subject to the immutability rule...and is subject to whatever
mutability state it had when it was inserted.)