Reimagine REFINEMENT! as a 2-element PATH!
This commit does a rethinking of the definition of a REFINMENT! such
that /word is the parallel to what would be [_ word] as a block.  This
opens up the possibilities for greater expressions of paths in the
pattern `/foo/bar/...`

Similar to how the QUOTED! change generalized what were previously
dedicated types, this adds REFINEMENT! as a parse rule and some
compatibility checks.  Unlike LIT-WORD? and LIT-PATH? however, the
matching of the pattern for REFINEMENT? and the term may be worth
keeping to idiomatically identify this path subclass.

As a first take, this is more inefficient than a refinement, as it
actually does allocate 2-element arrays for each usage (which can't
fit in 2 cell spaces, as it must account for an end as well, so it
requires a full 4 cell-sized array allocation, in addition to a
series node).  This makes a refinement cost roughly the overhead of
6 additional value cells to the one cell it occupied previously.

Next step will be to adapt the "in-situ" escaping technique (which was
made to put higher quoting levels into cells directly) such that it
can represent this form of 2-element path at nearly no added cost,
while still appearing to be a PATH!.