Reimagine REFINEMENT! as a 2-element PATH!
This commit does a rethinking of the definition of a REFINEMENT! such
that /word is the parallel to what would be [_ word] as a block.  This
opens up the possibilities for greater expressions of paths in the
pattern `/foo/bar/...`

Similar to how the QUOTED! change generalized what were previously
dedicated types, this adds REFINEMENT! as a parse rule and some
compatibility checks.  Unlike LIT-WORD? and LIT-PATH? however, the
matching of the pattern for REFINEMENT? and the term may be worth
keeping to idiomatically identify this path subclass.

As a first take, this is more inefficient than a refinement, as it
actually does allocate 2-element arrays for each usage (which can't
fit in 2 cell spaces, as it must account for an end as well, so it
requires a full 4 cell-sized array allocation, in addition to a
series node).  This makes a refinement cost roughly the overhead of
6 additional value cells to the one cell it occupied previously.

This commit also streamlines path scanning to use Scan_To_Stack()
At one point, scanning paths was simplified to ask a recursive call
to generate an array.  This changes to not generate an array, rather
to build the new array on the stack first...so that the decision to
make an array or not can be made when looking at what was scanned.

Near-term efficiency tricks should be able to easily make `/a` fit
into a single cell without an allocation.  Groundwork has been laid
to make arbitrary other values fit, such as `/1.0` or even `a/` to
lean the path the other way.  Longer term this should generalize to
more than paths so that immutable blocks and groups like `(a)` or
`[1.0]` can also fit into a single cell.