Make JOIN act like JOIN-OF (reverting to Rebol2 semantics)
Given Rebol's mutable-by-default semantics, it seemed for a time that
it might make sense to take the "mutable joining operation" (REPEND)
and make JOIN do that.

The idea was tried out for a time by having JOIN do this in the
mezzanine, but exporting the routine under the name ADJOIN.  Then
what had formerly been JOIN was done with JOIN-OF.

This particular avenue of thought didn't lead to any great new outcome.
Moreover, the conversion of paths to an immutable type means that the
system really does need a system-level way to do a JOIN that is not
thought of as "copy the thing and append to it" (as COPY on an
immutable path is a no-op, and APPEND will not accept it as a type).

So this goes ahead and reverts it--which is not a particularly complex
thing to do.  The mezzanine uses of JOIN XXX YYY are changed to:

     APPEND XXX REDUCE YYY

ADJOIN is deleted, and JOIN-OF is replaced with JOIN.