Fix buffer overrun in DELINE
DELINE operates in place with two pointers--one source and one dest,
because it knows it will always be making a shorter string.

The process was supposed to be bounded by the total character count
(as the string may contain embedded NULs in the current formulation).
So the count was compared to the length on every loop iteration.
However, advancement occuring inside the loop wasn't being accounted
for, which meant that it wouldn't realize it was reading past the
string boundary

This moves the bumping of the index inside the loop so it can be done
for both cases.