PARSE COLLECT patch for R3C branch
This is a patching in of the fairly experimental COLLECT and KEEP
functionality in PARSE.  It is distinct from Red's COLLECT in PARSE
due to using a name for the collection, distinct from the result:

    red>> parse "aaabb" [collect [some [keep "a"]] end]
    == [#"a" #"a" #"a"]

vs.

    r3c>> parse "aaabb" [collect data [some [keep "a"]]]
    == "bb"

    r3c>> data
    == ["a" "a" "a"]

This allows one to independently measure how far a parse gets vs.
the collection, and many collections can be done.

It also implements backtracking, so failed rules will remove their
additions from the collection:

   red> parse "ab" [collect [keep "a" keep "c" | keep "a" keep "b"]]
   == [#"a" #"a" #"b"]

vs.

   r3c> parse "ab" [collect data [keep "a" keep "c" | keep "a" keep "b"]]
   == ""

   r3c> data
   == ["a" "b"]

Other differences are that KEEP of string portions keeps strings and
not individual characters, which is the kind of behavior that could
be reviewed and discussed.

Maintenance of this feature may be difficult given the divergence of
the R3C code from mainline.  One feature that is not implemented is
the use of a GET-BLOCK! to keep material that did not originate from
PARSE or a match rule.