Fix bootstrap build issue and clean up Travis YAML
Bootstrap builds were failing on Travis but not reporting it.  This
corrects some issues that were causing breakages.

It also deletes the makefile before doing a second build on Travis
to keep this from happening in the future.  But in addition to that,
it vets the Travis file (now a product of many authors and years of
cutting and pasting from the Internet) for some basic bash practices,
along with comments explaining why those practices should be used.
This mostly involves avoiding superfluous delimiters and putting in
delimiters when they are actually needed.