Add SYM-BLOCK!, SYM-WORD!, SYM-GROUP!, SYM-PATH!
With enough base data types freed up to allow it, this adds the non
evaluating SYM-XXX! types.  They start with @, and SYM-WORD! is
specifically intended to pick up the slack from losing ISSUE! as an
ANY-WORD! type (returning it to being an ANY-STRING!)

In order to keep the ANY-INERT! check cheap, this makes some of the
tests for things like ANY-PATH! or ANY-BLOCK! more complicated.  They
are used more in the debug build than anything.  This doesn't worry
too much about the details of it, using the TS_ARRAY, TS_WORD, etc.
typesets generated by the build process vs. range comparisons on the
byte values.