Proof-of-Concept Extension types implementation (4 types)
Though incomplete, this is enough groundwork in extension types that
four types that were previously "built-in" (IMAGE!, GOB!, VECTOR!, and
STRUCT!) are now no longer in the %types.r file.  The basic mechanism
is established, where their cells give up the "Extra" slot and must
fit their payloads into the remaining two slots of the cell.

The EVENT! type is not changed to use this mechanism, as doing so would
mean one of its design points of fitting events entirely in a cell
would not be possible.  Hence it gets a reserved cell type and fills
in the table entries when the event extension loads.

Currently the notion of the custom type is captured by a node which is
a BINARY! containing dispatchers for the various "hooks" (molding,
generic actions, TO conversions, MAKE, etc.)  Only four slots are
reserved for storing canon versions of these type values at the moment,
but the vision is that the slots would be identified by URL!, and that
"stub" datatypes could exist prior to the extension being loaded.
This way parameter types could be made to refer to the types when the
extension isn't loaded yet.  The URL would ultimately belong in the
"type spec", along with any description.

Though this is not completely worked out, the state it is in offers
an advantage in recovering several non-extended type slots, which are
currently needed for the @ types.