Permissive GROUP! invisibility for R3C
With the rise of the role of "invisibles" in Ren-C, more and more
parameterized ones came to exist.  This began to make it compelling
to be able to put them in groups, and have the group itself dissolve
if its contents disappeared:

    >> 1 + 2 (comment "hi")
    == 3

    >> case [false [<no>] (elide print "heya") true [<yes>]]
    heya
    == <yes>

This patch brings a partial implementation of permissive group
invisibility to R3C.  Doing it truly "correctly" would require changes
that would be too difficult to patch in, but this covers a fair bit.

Also renames EVAL to REEVAL, as Ren-C's philosophy that abbreviated
forms should act as their full-named forms.  This forced EVAL to be
a synonym for EVALUATE.  Hence REEVAL is a synonym for REEVALUATE.
As no such routine exists in historical Rebol, this fix is considered
worthwhile.  It is integrated here as behavior changed to allow the
evaluation of invisibles to return a VOID!.