Require BLOCK! to modify an ANY-ARRAY! unless /ONLY
For aggregate values, there has always been an ambiguous situation in
Rebol as to whether an operation (e.g. APPEND, INSERT, CHANGE) means
to treat another aggregate value as a collection of things to splice
or a single item:

    ; Option (1)
    >> append [a b c] [d e]
    == [a b c d e]

    ; Option (2)
    >> append [a b c] [d e]
    == [a b c [d e]]

From a technical point of view, the more fundamental operation is (2)
That is because with iteration you could implement (1) by adding the
items one at a time.  But if blocks always spliced in these operations
and (1) was the only mechanic, you'd have no way to append a block to
a block.

Despite adding items atomically being more "fundamental", Rebol chose
to make most operations interpret aggregates as being spliced by
default.  To get the behavior in (2), an /ONLY refinement would have
to be used.

The benefits of this choice aren't obvious at first glance, when
thinking purely in mechanical terms.  But usage of the language in
practice reveals that there's leverage in this assumption.  It also
aligns well with string operations, which have no choice but to splice
and have no parallel to appending another string atomically, e.g.:

    >> append "abc" "de"
    == "abcde"  ; you don't want ["abc" "de"]

HOWEVER, accepting splice-by-default led to some poor invariants; as
non-aggregates would fall back to just adding a single item.  Code
that was meant to be generic could appear to work.  So imagine a
generic function like:

    block: []
    add-item: function [x] [append block x]

The intent to use /ONLY here could be very easily overlooked...
especially since code can appear to work for a time and then suddenly
not work when a BLOCK! is used (or historically also PATH!, GROUP!).
Calling it a newbie mistake is not true: it is a chronic problem in
the system due to the duality that aggregates are also values
themselves.

The solution has often been made to reverse the bias and require an
explicit /SPLICE.  However this carries its own set of tradeoffs,
making formerly elegant code inelegant.

THIS EXPERIMENTAL COMMIT EMBRACES THE HISTORICAL CONVENTION OF SPLICE IF
THE TARGET IS AN ANY-ARRAY!, BUT ONLY IF THE SOURCE MATERIAL IS BLOCK!,
AND WILL ERROR ON OTHER SOURCE DATA.  The idea is to train users and
ingrain it in their mind to think of using /ONLY whenever they intend
to work with single items within a block.  This elevates blocks to a
greater common currency in the system.

For those looking at callsites where the intent seems obvious, this may
appear to add a bit of hassle.  e.g. instead of `append block <tag>`
one must say `append block [<tag>]` or `append/only block <tag>`.
But the benefits are for `append block (black-box-expression)`, where
the poor invariants have been rearing their ugly head since the
beginning of the language...where the convenience that drove APPEND
to splice by default runs up against the intuition that those writing
generic code have about what the more fundamental operation is.  By
forcing the black box expression to resolve to a block when /ONLY is
not used, the callsite becomes more sensible.

To mitigate any perceived pain, "modal" parameters are used to allow
the /ONLY flag to be conveyed by a single character in many of the most
common cases:

    b1: [a b c]
    b2: [d e]

    >> append b1 @b2
    == [a b c [d e]]

    >> append b1 @(reverse copy b2)
    == [a b c [d e] [e d]]

    >> append b1 @[f g h]
    == [a b c [d e] [e d] [f g h]]

This commit also rethinks MAP-EACH to splice by default, adding an
/ONLY option to avoid that (which can be conveyed via a modal parameter
on the code block itself).  COLLECT is also given an /ONLY option, so
that /ONLY is removed from KEEP and it always means keep one item.