Tweak easy-ecc functions and add secp256r1 native
This adds keypair generation and secret-making for secp256r1 via
the "easy-ecc" C file.  The way that easy-ecc worked it used a
compressed form of key which discarded the Y coordinate of the pair
and turned it into an extra byte indicating the sign, so it would
be recalculated.  In order to better impedance match against the
TLS protocol, this incorporates natives with the uncompressed format,
so there are `_xy` forms exposed of the inner routines.

As a first step, this only does generation and shared secret calculation
but does not do verification.