TAKE of empty series returns NULL patch for R3C
Conservative policies about mutating operations led routines like
APPEND to not be silent when they were given NONE! values to append
to, even though read-only routines like SELECT were willing to let
a NONE! input slide.

Early attempts in Ren-C to be conservative thus built TAKE on top of
a lower-level TAKE* that was NULL-tolerant, but alerted you if the
result of the TAKE was nothing.

This seemed to be misguided and was reverted.  This patches allowance
of soft failure onto R3C.