ENBIN and DEBIN patch for R3C
This adds the ENBIN and DEBIN natives to R3C.  It does not do the
corresponding updates to actually use those natives in implementation
(e.g. of TLS) due to the large number of changes.  Hence the old
TO-INTEGER/UNSIGNED is left in that code as-is.

However, it makes them available for use.

https://forum.rebol.info/t/1270