Changes for R3C to disable warnings under new compilers
The RIGOROUS compilation mode offers useful build warnings.  But these
warnings are not part of the language standard, and new ones are
introduced with each compiler.

While Ren-C was kept in sync with warnings over time, trying to build
the older branch of R3C with modern compilers doesn't have those fixes.
This addresses the issues by either disabling the warnings or making
the change that was made on the Ren-C master branch.

In particular, Spectre-related warnings are ignored vs. mitigated, as
the patch was too complex to be worth it (considering it is unlikely
that a spectre-mitigated build of R3C would ever be built, and the
only effect of not having the patches would be being slightly slower).