R3 patch PRINT NEWLINE legal, prints 1 newline (not 2)
While PRINT in Ren-C was restricted to block and text input, it had
for a time a PRINT-NEWLINE native to work around this by writing a
single newline (not two) to stdout.

Looks-wise, it was decided that PRINT accepting the CHAR! of NEWLINE
and responding to it with one (and only one) newline output was more
legible and not a problem.  This backpatches R3C to allow this and
eliminates the PRINT-NEWLINE function.