Add /RESULT output for CATCH and FAIL, fix CATCH/QUIT
This patches R3C to be able to request a "result" in addition to the
standard caught/errored output.

    >> trap/result [fail "Error"] 'nonerror
    == make error! [...]

    >> type of :nonerror
    == void!

    >> trap/result [1 + 2] 'nonerror
    ; null

    >> nonerror
    == 3

This ability is leveraged in mainline to do multiple return values:

    [error nonerror]: trap [...]

R3C lacks this ability, but mainline is backwards compatible (goes in
order of the refinements in the spec as "outputs").  So the refinement
method can still be used.

Includes a small made for CATCH/QUIT in mainline: "The CATCH/QUIT was
not passing through non-quit throws, instead returning NULL."