Patch R3C to modern SET/GET conventions; add NON
The current mode of thinking in Ren-C is that SET, SET-WORD!, SET-PATH!
do not complain when assigned either NULL or VOID!.  This is based
on the premise that the main channel for catching typos or errors is
through WORD! access...which complains when a VOID! is referenced.
This catches typos on the GET side.

(Catching typos on the set side with values isn't really possible,
as `mispeled-var: 100` involves no void!s or NULLs anyway)

This brings R3C up to date with these ideas, including ones discussed
today in chat.  That is to make SET have parity with SET-WORD! and
not require any /OPT or /ANY refinement.  Instead, it's up to users
to check the value they are assigning.  Examples would be:

    set var non void! (...some expression...)
    set var non null (...some expression...)
    set numeric-var ensure [integer! decimal!] (...some expression...)

Encouraging this direction is far more general, and can be a pattern
that those interested in it can use (or make their own SET or SET
variants that fold these checks into them).

On the GET side however, the importance of leveraging VOID! for its
main purpose as a way to avoid typos suggests being more prescriptive.
This canonizes the refinement to GET to allow VOID! as the historical
/ANY, which makes actual sense as it can be read to mean ANY-VALUE!
(including VOID!) is legal.

The SET-WORD! and SET-PATH! cases are modified to permit VOID!
assignments as well.

This is part one of a two part commit, as for this to be workable the
erroring state for "ornery" variables (typos, etc.) has to be VOID!
and not NULL.