"ornery" values void (uninitialized locals, typos)
The initial pursuit of a "true unset" originally conflated a typo or
otherwise intentionally "ornery" variable with the complete absence
of a value.  Over time this idea broke down as NULL emerged as the
"true none" as opposed to the true mean "UNSET!" value.  So the latter
was reintroduced as VOID!.

This implements the mainline notion that VOID! is used as the state for
uninitialized locals and typos, and NULL is reserved for the true
equivalent to C's NULL, as not pointing to any value cell and not
being able to be put into a BLOCK!.

Following on the previous commit, this helps by eliminating UNSET the
native...as one can set variables to NULL or VOID! with SET-WORD! or
SET without error.  Hence `unset 'var` can be changed into `var: null`
or `var: void` depending on what the actual intent was.