Make PAIR! preserve memory of INTEGER!/DECIMAL! x and y
R3-Alpha used C `float` numbers in PAIR!, in order to compactly fit
two floating point numbers into a single cell.  (It would not be
possible to fit two 64-bit integers into a single cell as that would
use the full space and leave none for the header even saying it
was a PAIR!...similarly it could not fit two 64-bit doubles.)

This made a new numeric range and type to deal with that was not in
use for INTEGER! or DECIMAL!.  Ren-C switched to a new trick that
made a generalized compact method for storing two of ANY-VALUE! in
a single series node, through what was known as a "pairing".  This
helped eliminate a lot of `float`-specific math.

Long-term the goal was to see if this could be used more generally
as pairs of any type.  But since PAIR! was not a particularly high
priority this was not pursued.  So it converted any inputs (even
ones that started INTEGER!) into a DECIMAL! in the pair.  This gave
compatible behavior to R3-Alpha, e.g.

   r3-alpha>> p: 2x4

   r3-alpha>> p/1
   == 2.0

This commit implements a concept of preserving the type, so that a
PAIR! which is made with x or y as INTEGER! or DECIMAL! will preserve
which it was.  It also uses a frame reinvocation trick to make more
operations work pairwise:

   >> p: 5.5x10

   >> p + 2
   == 7.5x12

   >> p + 3x4
   == 10.5x16

   >> type of p/1
   == #[datatype! decimal!]

   >> type off p/2
   == #[dataatype! integer!]

The tests are not adapted, as it is being left to @rgchris to spec the
most desirable behaviors for PAIR!.