Rename "new" APPLY to APPLIQUE
When Ren-C first introduced a generic APPLY primitive, it used the
idea of a BLOCK! of code that would name each parameter explicitly.
It did not use positional arguments.

This "new APPLY" isn't as general as the one currently imagined, which
would blend the best of positional apply with named-arguments.  That
APPLY has not been written, and will require "refinements are their
arguments" to work cleanly...which is not in R3C.

What mainline has done is to move "new APPLY" to the otherwise not very
desirable name APPLIQUE.  Whether this is a good name or not, it's not
going to be used for anything else.  The real APPLY will be coming
down the pipe.