GET-WORD! errors on VOID!, only suppresses ACTION!
(similar patch applied to master branch)

This change aligns Ren-C with Rebol2, in considering the VOID! (e.g.
"UNSET!") state to be an error when fetched via GET-WORD!.  That means
that GET/ANY must be used to read VOID!.

The motivation for this is to make sure that when people use GET-WORD!
to suppress function evaluation, they aren't also getting something
they may not want...e.g. the suppression of errors on typos if the
word looks up to void.

As a compromise for performance, GET* is added to mean "GET/ANY" as
a native.  This makes `get* 'var` a bit closer to the performance
profile of `:var` because there is no path processing...and it is
a native so it does not incur specialization overhead.