Remove direct Scan_Tuple use from DNS extension
The DNS extension is currently using very low-level routines in
general...and it would probably be better if it were more abstract.
In any case, it calls "Scan_Tuple" directly...so this changes it to
use the API and TRANSCODE.  Performance of this extremely rare
tuple conversion is not that important compared to reducing the
amount and complexity of the source...and testing more API codepaths.