First-Cut Implementation of Tuples as Arrays (WIP)
This begins the transition to generic tuples by removing the assumption
that tuples are byte arrays, leaving that optimization to lower levels
of the implementation.  By factoring it in this way, the same
optimizations can be shared between PATH! and TUPLE!.

As with PATH!, this requires changing the accessors to use a TUPLE_AT
with an index, instead of accessing the internals automatically.
Wrappers are added to give compatibility with the byte-oriented tuple
initialization and extraction.

This is a work-in-progress, which needs too be merged with path
implementation changes before going too far with testing.