Change `<...>` to `<variadic>`
There was already some question about `<...>` being used to express
variadic parameters, as `...` is slated to mean "TBD" when used in
code.  However, new rules would make that a TUPLE! with `<` at the
head and `>` at the tail...not a TAG!.

This doesn't technically preclude it from being used for its visual
properties in a dialect (like the function spec dialect).  But it
messes with the current implementation so switching to `<variadic>`
is the easiest choice for the moment.  Review if other notations
(e.g. putting type specs in a "doubled-block" would be better).

In order to avoid making clients change this twice, support for
the TUPLE!-based `<...>` is added as an interim measure before a
final syntax for variadics is decided on.