Make tuples load in path (but not vice versa), fixes
This triggers a nested scan so that `a.b/c` knows to act as `(a.b)/c`
and not `a.(b/c)` or `a.b /c` or `a. b/c`, etc.  This basically
involves knowing to trigger a recursion on seeing tuples while path
scanning, but to go up a stack level to continue path scanning if
a slash is seen while scanning a tuple.

In order to get the scan of `.` to work, this goes ahead and makes a
-dot-0- analogue to -path-0- word, and generalizes that feature.