"Outdent" Scan_To_Stack()
Scan_To_Stack() is a long function, which is indented a level to
account for a `while(true)`.  The function is so long that `goto` is
clearer than continue or break, so it might as well be a labeled
block of code.

(With significant rewriting of the path and tuple scanning, this is
as good a time to change it as any.)