Make scanner exit condition clearer
There had been a check for TOKEN_END after the Locate_Token() to avoid
validating the ss->begin < ss->end, but also an instance in the switch.
By eliminating the unnecessary switch case, the termination condition
can be made clearer, as well as its error state when a BLOCK! or
GROUP! didn't find its terminal character.