Booting implementation of refinement optimization
This is a first crack at saying that a value with its "MIRROR_BYTE"
set to REB_WORD but its "KIND_BYTE" set to some form of ANY-PATH!
will act as a refinement (so long as it is not the special words
`-slash-1-` or `-dot-1`)

This allows refinements to be stored in a single cell, as in R3-Alpha.
However, to be seen "AS" an array they must currently be converted.
This is done implicitly by PARSE on an INTO, so all refinements in
function specs processed by usermode were currently making temporary
2-element arrays.

Here that changes to use the REFINEMENT! type constraint.