Make GET-GROUP! behave same as GROUP!
The mechanism for making GET-GROUP! re-dispatch mutated the "heart"
bytes of cells, which is no longer safe.  All such instances need
to be patched, but this case is now being looked at as a way to signal
branching constructs not to voidify a null branch.  Since there is
little chance that :(x) is going to be used to mean simply GET X
at this point, the behavior is unified with GROUP!.

`(x):` remains open as to what it will be used for.