TO TUPLE! of TEXT experiment
The TO and MAKE conversions are historically ill-defined in Redbol.
This experimentally builds TO TUPLE! of TEXT! on top of AS applied to
LOAD, so that it tolerates forms like "1 2 3" as well as "1.2.3".

It's not efficient, but it's good enough to make HTTP reads work and
it a small patch that exercises existing code.